starts refactor into Resources component
