+++ apps/gp2-frontend/src/working-groups/WorkingGroupDetail.tsx
@@ -4,7 +4,7 @@ import { Frame, useBackHref } from '@asap-hub/frontend-utils';
 import {
   WorkingGroupDetailPage,
   WorkingGroupOverview,
-  WorkingGroupResources,
+  Resources,
 } from '@asap-hub/gp2-components';
 import { NotFoundPage } from '@asap-hub/react-components';
 import { Redirect, Route, Switch } from 'react-router-dom';
@@ -49,7 +49,7 @@ const WorkingGroupDetail = () => {
               }
             >
               <Frame title="Resources">
-                <WorkingGroupResources {...workingGroupData} />
+                <Resources resources={workingGroupData.resources} />
               </Frame>
             </Route>
           )}
