+++ packages/gp2-components/src/templates/__tests__/Resources.test.tsx
@@ -1,10 +1,10 @@
 import { gp2 } from '@asap-hub/model';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import WorkingGroupResources from '../WorkingGroupResources';
+import Resources from '../Resources';
 
-describe('WorkingGroupResources', () => {
-  const getResources = (length = 1): gp2.WorkingGroupResource[] =>
+describe('Resources', () => {
+  const getResources = (length = 1): gp2.Resource[] =>
     Array.from({ length }, (_, itemIndex) => ({
       type: 'Link' as const,
       title: `resource title ${itemIndex}`,
@@ -16,7 +16,7 @@ describe('WorkingGroupResources', () => {
   };
 
   it('renders heading', () => {
-    render(<WorkingGroupResources {...defaultProps}></WorkingGroupResources>);
+    render(<Resources {...defaultProps} />);
     expect(
       screen.getByRole('heading', { name: /Resource List/i }),
     ).toBeVisible();
@@ -24,12 +24,7 @@ describe('WorkingGroupResources', () => {
   it('renders a resource title', () => {
     const [resource] = getResources();
     resource.title = 'resource title';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource]} />);
     expect(
       screen.getByRole('heading', { name: /resource title/i }),
     ).toBeVisible();
@@ -37,12 +32,7 @@ describe('WorkingGroupResources', () => {
   it('renders a resource description', () => {
     const [resource] = getResources();
     resource.description = 'resource description';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource]} />);
     expect(screen.getByText(/resource description/i)).toBeVisible();
   });
   it('renders a link resource external link', () => {
@@ -52,12 +42,7 @@ describe('WorkingGroupResources', () => {
       type: 'Link',
       externalLink: 'http://a-link',
     };
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resourceLink]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resourceLink]} />);
     expect(
       screen.getByRole('link', { name: /external link/i }),
     ).toHaveAttribute('href', 'http://a-link');
@@ -66,36 +51,21 @@ describe('WorkingGroupResources', () => {
   it('renders a link resource pill for a link', () => {
     const [resource] = getResources();
     resource.type = 'Link';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource]} />);
     expect(screen.getByText('Link')).toBeVisible();
   });
 
   it('should not render a note resource external link', () => {
     const [resource] = getResources();
     resource.type = 'Note';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource]} />);
     expect(screen.queryByTestId('external-link-0')).not.toBeInTheDocument();
   });
 
   it('renders a note resource pill for a note', () => {
     const [resource] = getResources();
     resource.type = 'Note';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource]} />);
     expect(screen.getByText('Note')).toBeVisible();
   });
 
@@ -103,12 +73,7 @@ describe('WorkingGroupResources', () => {
     const [resource1, resource2] = getResources(2);
     resource1.title = 'resource title 1';
     resource2.title = 'resource title 2';
-    render(
-      <WorkingGroupResources
-        {...defaultProps}
-        resources={[resource1, resource2]}
-      ></WorkingGroupResources>,
-    );
+    render(<Resources {...defaultProps} resources={[resource1, resource2]} />);
     expect(
       screen.getByRole('heading', { name: /resource title 1/i }),
     ).toBeVisible();
@@ -120,14 +85,14 @@ describe('WorkingGroupResources', () => {
   it('Renders show more button for more than 3 milestones', async () => {
     const resources = getResources(4);
 
-    render(<WorkingGroupResources resources={resources} />);
+    render(<Resources resources={resources} />);
 
     expect(screen.getByRole('button', { name: /Show more/i })).toBeVisible();
   });
   it('Renders show less button when the show more button is clicked', async () => {
     const resources = getResources(4);
 
-    render(<WorkingGroupResources resources={resources} />);
+    render(<Resources resources={resources} />);
 
     const button = screen.getByRole('button', { name: /Show more/i });
     userEvent.click(button);
@@ -136,7 +101,7 @@ describe('WorkingGroupResources', () => {
   it('does not show a more button for less than 3 milestones', async () => {
     const resources = getResources(3);
 
-    render(<WorkingGroupResources resources={resources} />);
+    render(<Resources resources={resources} />);
 
     expect(
       screen.queryByRole('button', { name: /Show more/i }),
@@ -145,7 +110,7 @@ describe('WorkingGroupResources', () => {
   it('displays the hidden milestones if the button is clicked', () => {
     const resources = getResources(4);
 
-    render(<WorkingGroupResources resources={resources} />);
+    render(<Resources resources={resources} />);
     expect(
       screen.getByRole('heading', { name: 'resource title 2' }),
     ).toBeInTheDocument();
