+++ packages/model/src/gp2/common.ts
@@ -1,0 +1,9 @@
+const resourceType = ['Link', 'Note'] as const;
+export type ResourceType = typeof resourceType[number];
+
+export interface Resource {
+  type: ResourceType;
+  title: string;
+  description?: string;
+  externalLink?: string;
+}
