+++ packages/model/src/gp2/index.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export * from './project';
 export * from './user';
 export * from './working-group';
