+++ packages/model/src/gp2/working-group.ts
@@ -1,4 +1,5 @@
 import { ListResponse } from '../common';
+import { Resource } from './common';
 
 const workingGroupMemberRole = [
   'Lead',
@@ -37,7 +38,7 @@ export type WorkingGroupDataObject = {
   secondaryEmail?: string;
   leadingMembers?: string;
   members: WorkingGroupMember[];
-  resources?: WorkingGroupResource[];
+  resources?: Resource[];
 };
 
 export type ListWorkingGroupDataObject = ListResponse<WorkingGroupDataObject>;
